config server update

+++ lunettesMS/src/main/java/com/esprit/lunettes/LunetteRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface LunetteRepository extends JpaRepository<Lunette, Integer> {
+public interface LunetteRepository extends JpaRepository<Lunette, Float> {
 
 	@Query("select c from Lunette c where c.marque like :marque")
 	public Page<Lunette> lunetteByMarque(@Param("marque") String n, Pageable pageable);
